Build a K-nearest neighbor (KNN) classifier

buildKnn loads data from two Andromeda tables, and inserts them into a KNN classifier.

buildKnn(outcomes, covariates, indexFolder, overwrite = TRUE)

Arguments

outcomes

An Andromeda table containing the outcomes with predefined columns (see below).

covariates

An Andromeda table containing the covariates with predefined columns (see below).

indexFolder

Path to a local folder where the KNN classifier index can be stored.

overwrite

Automatically overwrite if an index already exists?

Value

Nothing

Details

These columns are expected in the outcome object:

rowId	(integer)	Row ID is used to link multiple covariates (x) to a single outcome (y)

y	(real)	The outcome variable

These columns are expected in the covariates object:

rowId	(integer)	Row ID is used to link multiple covariates (x) to a single outcome (y)

covariateId	(integer)	A numeric identifier of a covariate

covariateValue	(real)	The value of the specified covariate