buildKnnFromPlpData {BigKnn}	R Documentation

Build a K-nearest neighbor (KNN) classifier from a plpData object

Description

Build a K-nearest neighbor (KNN) classifier from a plpData object

Usage

buildKnnFromPlpData(
  plpData,
  population,
  indexFolder,
  overwrite = TRUE,
  cohortId = NULL,
  outcomeId = NULL
)

Arguments

plpData
An object of type plpData.

population
The population.

indexFolder
Path to a local folder where the KNN classifier index can be stored.

overwrite
Automatically overwrite if an index already exists?

cohortId
The ID of the specific cohort for which to fit a model.

outcomeId
The ID of the specific outcome for which to fit a model.

[Package BigKnn version 1.0.1 Index]